feat: create documentation

+++ src/product/product.service.ts
@@ -1,13 +1,17 @@
+import { AuthService } from './../auth/auth.service';
+import { LocalStrategy } from './../auth/local.strategy';
 import { ProductDTO } from '../product/product.dto';
 import { ProductModel } from '../product/product.model';
-import { Injectable, NotFoundException } from "@nestjs/common";
+import { HttpCode, Injectable, NotFoundException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from 'typeorm';
 
 @Injectable()
 export class ProductService {
 
-    constructor(@InjectRepository(ProductModel) private productRepository: Repository<ProductModel>) {}
+    constructor(
+        @InjectRepository(ProductModel) private productRepository: Repository<ProductModel>,
+    ) {}
 
     async create(product: ProductDTO): Promise<ProductModel> {
         return await this.productRepository.save(product)
@@ -27,11 +31,24 @@ export class ProductService {
         return product
     }
 
-    async update(id: number, product: ProductDTO) {
-        return await this.productRepository.update(id, product)
+    async update(id: number, product: ProductDTO): Promise<ProductModel> {
+        const productFound = await this.productRepository.findOne(id)
+        
+        if(!productFound) {
+            throw new NotFoundException(`Id ${id} doesn't exists`)
+        }
+
+        await this.productRepository.update(id, product)
+        return await this.productRepository.findOne(id)
     }
 
-    delete(id: number): void {
+    async delete(id: number): Promise<void> {
+        const product = await this.productRepository.findOne(id)
+
+        if(!product) {
+            throw new NotFoundException(`Id ${id} doesn't exists`)
+        }
+
         this.productRepository.delete(id)
     }
